Queriy/requirements.txt
google-generativeai 
flask 
bcrypt
SQLAlchemy
